Fixed  errors on country

+++ Code/ftxui-starter/src/Country/Country.h
@@ -68,23 +68,23 @@ class Country {
     Alliance * getAlliance() {return this->allies;}
     double getMorale();
     double getEconomy();
-    double getPopulation();
+    int getPopulation();
     double getResearch();
     double getGoalRating();
     double getAggressiveness();
     double getResources();
-    double getNumSpies();
+    int getNumSpies();
     int getNumTroops();
     int getNumVehicles();
     int getNumEnemyRegions();
     void setMorale(double morale);
     void setEconomy(double economy);
-    void setPopulation(double population);
+    void setPopulation(int population);
     void setResearch(double research);
     void setGoalRating(double goalRating);
     void setAggressiveness(double aggressiveness);
     void setResources(double resources);
-    void setNumSpies(double numSpies);
+    void setNumSpies(int numSpies);
     void setNumTroops(int numTroops);
     void setNumVehicles(int numVehicles);
     void setNumEnemyRegions(int numEnemyRegions);
